% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{crnhughop}
\alias{crnhughop}
\title{CRN Hug + CRN Hop 
Two-scale coupling for Hop: 
  1. apply rejection-sampled maximal proposal coupling when fall under some threshold of squared distance
 2. apply GCRN when fall above the threshold}
\usage{
crnhughop(x0, y0, Time, Bounces, lam, kap, thresh, iter, logpi, gradlogpi)
}
\description{
CRN Hug + CRN Hop 
Two-scale coupling for Hop: 
  1. apply rejection-sampled maximal proposal coupling when fall under some threshold of squared distance
 2. apply GCRN when fall above the threshold
}
