% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{RWM2targets_optimal}
\alias{RWM2targets_optimal}
\title{Optimal coupling
#'@export}
\usage{
RWM2targets_optimal(x, y, h, iter, logpix, gradlogpix, logpiy, gradlogpiy)
}
\description{
Optimal coupling
#'@export
}
