% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{RWM2targets_refl}
\alias{RWM2targets_refl}
\title{Reflection coupling}
\usage{
RWM2targets_refl(
  x,
  y,
  h,
  iter,
  logpix,
  gradlogpix = NULL,
  logpiy,
  gradlogpiy = NULL
)
}
\description{
Reflection coupling
}
