\name{rwmcouplings-package}
\alias{rwmcouplings-package}
\alias{rwmcouplings}
\docType{package}
\title{
\packageTitle{rwmcouplings}
}
\description{
\packageDescription{rwmcouplings}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rwmcouplings}
\packageIndices{rwmcouplings}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rwmcouplings}

Maintainer: \packageMaintainer{rwmcouplings}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
