% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SampleLatentVariables}
\alias{SampleLatentVariables}
\title{Sample the latent variables from the model}
\usage{
SampleLatentVariables(T, sig, phi)
}
\description{
Sample the latent variables from the model
}
