% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svmLaplaceCRNRWM}
\alias{svmLaplaceCRNRWM}
\title{CRN-coupled RWM algorithms targeting SVM and a Laplace approximation}
\usage{
svmLaplaceCRNRWM(
  x0,
  y0,
  y_data,
  beta,
  sigma,
  phi,
  mu,
  Omega,
  Omega_chol,
  h,
  iter,
  thin
)
}
\description{
CRN-coupled RWM algorithms targeting SVM and a Laplace approximation
}
