% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svmLaplaceGCReflRWM}
\alias{svmLaplaceGCReflRWM}
\title{Corrected reflection}
\usage{
svmLaplaceGCReflRWM(
  x0,
  y0,
  y_data,
  beta,
  sigma,
  phi,
  mu,
  Omega,
  Omega_chol,
  h,
  iter,
  thin
)
}
\description{
Corrected reflection
}
