% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svmLaplaceGCRNRWM}
\alias{svmLaplaceGCRNRWM}
\title{GCRN-coupled RWM algorithms targeting SVM and a Laplace approximation}
\usage{
svmLaplaceGCRNRWM(
  x0,
  y0,
  y_data,
  beta,
  sigma,
  phi,
  mu,
  Omega,
  Omega_chol,
  h,
  iter,
  thin
)
}
\description{
GCRN-coupled RWM algorithms targeting SVM and a Laplace approximation
}
