% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svmLaplaceReflRWM}
\alias{svmLaplaceReflRWM}
\title{Reflection-coupled RWM algorithms targeting SVM and a Laplace approximation}
\usage{
svmLaplaceReflRWM(
  x0,
  y0,
  y_data,
  beta,
  sigma,
  phi,
  mu,
  Omega,
  Omega_chol,
  h,
  iter,
  thin
)
}
\description{
Reflection-coupled RWM algorithms targeting SVM and a Laplace approximation
}
