% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svmLaplacetwoscaleGCReflRWM}
\alias{svmLaplacetwoscaleGCReflRWM}
\title{Two-scale corrected reflection}
\usage{
svmLaplacetwoscaleGCReflRWM(
  x0,
  y0,
  y_data,
  beta,
  sigma,
  phi,
  mu,
  Omega,
  Omega_chol,
  h,
  iter,
  thin,
  thresh
)
}
\description{
Two-scale corrected reflection
}
