% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svmReflMaxRWM}
\alias{svmReflMaxRWM}
\title{RWM algorithms, with reflection-maximal coupling}
\usage{
svmReflMaxRWM(x0, y0, y_data, beta, sigma, phi, h, iter, thin)
}
\description{
RWM algorithms, with reflection-maximal coupling
}
