% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svmTwoScaleGCReflRWM}
\alias{svmTwoScaleGCReflRWM}
\title{Two-scale coupling, with corrected reflection and reflection.}
\usage{
svmTwoScaleGCReflRWM(x0, y0, y_data, beta, sigma, phi, h, iter, thin, thresh)
}
\description{
Two-scale coupling, with corrected reflection and reflection.
}
