% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svmTwoScaleGCRNRWM}
\alias{svmTwoScaleGCRNRWM}
\title{RWM algorithms, with two-scale coupling using GCRN when (|x-y|^2 >= thresh) and using reflection-maximal coupling when (|x-y|^2 < thresh)}
\usage{
svmTwoScaleGCRNRWM(x0, y0, y_data, beta, sigma, phi, h, iter, thin, thresh)
}
\description{
RWM algorithms, with two-scale coupling using GCRN when (|x-y|^2 >= thresh) and using reflection-maximal coupling when (|x-y|^2 < thresh)
}
