% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{twoscalehmc}
\alias{twoscalehmc}
\title{two-scale GNRN-RWM + CRN-HMC, mixed}
\usage{
twoscalehmc(
  x0,
  y0,
  Time,
  L,
  gamma = 0.05,
  h,
  thresh = 0.1,
  iter,
  logpi,
  gradlogpi
)
}
\description{
two-scale GNRN-RWM + CRN-HMC, mixed
}
